// src/core/store/ducks/user/user.types.ts
enum ActionTypes {
  REGISTER_PENDING = "user/REGISTER_PENDING",
  REGISTER_SUCCESS = "user/REGISTER_SUCCESS",
  REGISTER_REJECT = "user/REGISTER_REJECT",
  REGISTER_ERROR = "user/REGISTER_ERRROR",
  LOGIN_PENDING = "user/LOGIN_PENDING",
  LOGIN_SUCCESS = "user/LOGIN_SUCCESS",
  LOGIN_REJECT = "user/LOGIN_REJECT",
  LOGIN_ERROR = "user/LOGIN_ERROR",
  LOGOUT = "user/LOGOUT"
}

export default ActionTypes;